move data classes that are only used by BinaryData to that file
