+++ simulariumio/data_objects/frame_data.py
@@ -1,4 +1,4 @@
-from typing import Dict, Union, Tuple
+from typing import Dict, Union
 
 
 class FrameData:
@@ -24,56 +24,3 @@ class FrameData:
         self.n_agents = n_agents
         self.time = time
         self.data = data
-
-
-class DataIndices:
-    def __init__(self, offset: int, length: int):
-        """
-        This object holds offset and length data for a binary
-        simularium data block
-
-        Parameters
-        ----------
-        offset : int
-            Number of bytes the block is offset from the start of the byte array
-        length : int
-            Number of bytes in the block
-        """
-        self.offset = offset
-        self.length = length
-
-    def get_start_end_indices(self) -> Tuple[int, int]:
-        """
-        Return the start and end indicies for the data block
-        """
-        end = self.offset + self.length
-        return self.offset, end
-
-
-class FrameMetadata:
-    def __init__(self, offset: int, length: int, frame_number: int, time: float):
-        """
-        This object holds metadata for a single frame of simularium data
-
-        Parameters
-        ----------
-        offset : int
-            Number of bytes the block is offset from the start of the byte array
-        length : int
-            Number of bytes in the block
-        frame_number : int
-            Index of frame in the simulation
-        time : float
-            Elapsed simulation time of the frame
-        """
-        self.offset = offset
-        self.length = length
-        self.frame_number = frame_number
-        self.time = time
-
-    def get_start_end_indices(self) -> Tuple[int, int]:
-        """
-        Return the start and end indicies for the frame
-        """
-        end = self.offset + self.length
-        return self.offset, end
